stream-load/requirements.txt
google-cloud-bigquery
google-cloud-functions
pytest==5.4.3

